% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVDJ.R
\name{addVDJ}
\alias{addVDJ}
\title{Add 10x VDJ Data to Seurat Obj}
\usage{
addVDJ(seurat_obj, type_v, dir_v)
}
\arguments{
\item{seurat_obj}{gene expression seurat object}

\item{type_v}{either "t" for TCR or "b" for BCR}

\item{dir_v}{directory where "filtered_contig_annotations.csv" and "clonotypes.csv" are found.}
}
\value{
Exact same seurat object with new metadata columns with clonotype information: 
[type_v]_clonotype_id and [type_v]_cdr3s_aa
}
\description{
Add 10x VDJ T or B data to an initialized GEX Seurat object. Modified from function found:
https://ucdavis-bioinformatics-training.github.io/2020-Advanced_Single_Cell_RNA_Seq/data_analysis/VDJ_Analysis_fixed
}
