% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedBar.R
\name{annotatedBar}
\alias{annotatedBar}
\title{Annotated Bar Plot}
\usage{
annotatedBar(
  data_dt,
  x_v,
  stat_v = "count",
  position_v = "stack",
  fill_v,
  fillColors_v = NULL,
  title_v = NULL,
  annot_lsv = NULL,
  annotColors_lsv,
  testTime_v = F
)
}
\arguments{
\item{data_dt}{melted data table for plotting}

\item{x_v}{name of x-axis variable (usually Sample_ID, Treatment, etc.). Must be column in data_dt}

\item{stat_v}{argument to 'stat' parameter of geom_bar. Haven't played around with having this be anything but 'count'}

\item{position_v}{argument to 'position' parameter of geom_bar. 'stack' (default) is counts; 'fill' turns to percentage by filling out of 1.}

\item{fill_v}{name of fill variable. Must be column in data_dt}

\item{fillColors_v}{optional named color vector. Values are colors, names are values of data_dt[[fill_v]]}

\item{title_v}{optional plot title}

\item{annot_lsv}{list of annotations to add below the plot. e.g. annot_lsv = list("outName1" = "colName1", "outName2" = "colName2")}

\item{annotColors_lsv}{list of colors for annotations. list element names are same as annot_lsv names; list elements are named color vectors, names are values of annot_lsv colNames in data_dt}
}
\value{
list of gg objects output by ggarrange. Combo plot, data only, annotations only.
}
\description{
Standard bar-plot with heatmap-like x-axis annotations
}
\details{
make a standard ggplot bar plot with extra annotations
}
