% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcENS.R
\name{calcENS}
\alias{calcENS}
\title{Calc Effective Number of Species}
\usage{
calcENS(values_v, proportion_v = T)
}
\arguments{
\item{values_v}{vector of values to compute with.}

\item{proportion_v}{logical indicating if values are proportions or actual values}
}
\value{
inverse sum of squared proportions of values_v
}
\description{
Use the proportions of populations to calculate effective number of species. This is a general function, but I use
it mainly to check the maximum iLISI when comparing harmony integrations.
Simpson Diversity Index (D) = sum(prop_i)^2, where prop_i = ni / N, where ni = abundance of species i and N = total sample number
Inverse = 1/D
}
