% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcErrorBars.R
\name{calcErrorBars}
\alias{calcErrorBars}
\title{Calculate Error Bars}
\usage{
calcErrorBars(
  data_dt,
  kind_v = "MSEM",
  col_v = "value",
  groupCol_v,
  fCol_v = NA,
  fVal_v = NA
)
}
\arguments{
\item{data_dt}{input data.table (most likely melted and ready for ggplot)}

\item{kind_v}{which bars to make. MSEM = mean +/- SEM; MSD = mean +/- sd; MQ = median and lower/upper quartiles}

\item{col_v}{which column in data_dt is the measurement that is used.}

\item{groupCol_v}{which column to group calculations by. Treatment is common}

\item{fCol_v}{which column to further group calcs by. Used if faceting the output.}

\item{fVal_v}{values of fCol_v to group by.}
}
\description{
Calculate error bars for given measurement column in data_dt.
}
\details{
Will calculate center and upper/lower bounds of disribution by one of two methods 
and add columns to data_dt labeled "Lower", "Middle", and "Upper". Will group calculated
error bars by a grouping column and can further subdivide by a second column.
}
