% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkExcel.R
\name{checkExcel}
\alias{checkExcel}
\title{Check Excel}
\usage{
checkExcel(dir_v = NULL, wkbk1_v, wkbk2_v, outDir_v = NULL)
}
\arguments{
\item{dir_v}{optional path to the directory that wkbk1 and wkbk2 are in}

\item{wkbk1_v}{If dir_v == NULL, path to first workbook. If dir_v is set, then file name within dir_v that points to first workbook}

\item{wkbk2_v}{If dir_v == NULL, path to second workbook. If dir_v is set, then file name within dir_v that points to second workbook}

\item{outDir_v}{optional directory to write output if any is created. Default, NULL, will write to working directory.}
}
\value{
Returns any different worksheets
}
\description{
Load two excel workbooks and check them for equivalence
}
\details{
Read in all worksheets in each workbook and check them for equivalence
}
