% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSamples.R
\name{checkSamples}
\alias{checkSamples}
\title{Check Samples}
\usage{
checkSamples(dt1_dt, dt2_dt, col1_v = "Sample", col2_v = col1_v)
}
\arguments{
\item{dt1_dt}{first data.table}

\item{dt2_dt}{second data.table}

\item{col1_v}{column in first data.table to check. Default is 'Sample'}

\item{col2_v}{column in second data.table to check. Default is to be same as col1_v.}
}
\description{
Check two data.tables to see if they have different samples.
Could be any column in theory.
}
