% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMouseHumanGenes.R
\name{convertMouseHumanGenes}
\alias{convertMouseHumanGenes}
\title{Convert Gene Names}
\usage{
convertMouseHumanGenes(
  genes_v,
  species_v = "mmu",
  host_v = "https://dec2021.archive.ensembl.org/"
)
}
\arguments{
\item{genes_v}{vector of genes to convert}

\item{species_v}{species identity to convert FROM. If "mouse" or "mmu", will be converted to human.
If "human" or "hg", will be converted to mouse.}
}
\value{
genes_v with updated names
}
\description{
Convert gene names between mouse and human using ensembl database (adapted from Rossin Erbe)

TO DO - add either verbose option or multiple output option to also output the genes that didn't get converted.
original host: "https://dec2021.archive.ensembl.org/" this worked at first, but keep on getting connection time outs. The biomart
functions should check other mirrors, but can also try "https://www.ensembl.org".
}
