% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCutOff.R
\name{correlationCutOff}
\alias{correlationCutOff}
\title{Correlation Cut Off}
\usage{
correlationCutOff(
  data_dt,
  name_v = NULL,
  corrCol_v = "Corr",
  cutType_v = "kneedle",
  corrType_v = "Pearson"
)
}
\arguments{
\item{data_dt}{data.table of correlations. Rows = genes; columns = corr}

\item{name_v}{optional vector to be prepended to title.}

\item{corrCol_v}{column name that holds correlation values. Default is "Corr"}

\item{cutType_v}{vector indicating which method to use. kneedle (default), "arrow", or "both"}

\item{corrType_v}{should always be 'Pearson', adding as an argument in case you want to run this on different corr results (just used for plot titles)}
}
\value{
list of 3 plots and data.tables with results
}
\description{
Find appropriate rank and value for cut-off of correlation distribution

Use "elbow" methods to find a non-arbitrary cut-off for a set of correlation values
}
