% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachAllPackages.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{Clear loaded packages}
\usage{
detachAllPackages(unload_v = F)
}
\arguments{
\item{unload}{argument to detach. See ?detach for explanation.}
}
\value{
None. Detaches loaded packages.
}
\description{
Clear all loaded packages (excluding base). Useful for determining if a script has the appropriate packages loaded.
}
\examples{
detachAllPackages()
}
