% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineBreaks.R
\name{determineBreaks}
\alias{determineBreaks}
\title{Determine Breaks}
\usage{
determineBreaks(data_mat, colors_v = rev(colorRampPalette(brewer.pal(11,
  "RdBu"))(100)), cut_v = "max", print_v = F, verbose_v = T)
}
\arguments{
\item{data_mat}{matrix of counts. Most likely will be row-scaled, but doesn't have to be. if data.table provided, 1st column is assumed to be
an identifier column and will be assigned to row.names.}

\item{colors_v}{Vector of colors that will be used. Default is red (high) to blue (low)}

\item{cut_v}{Either a numeric value indicating which percentile to use (3 would use the 3rd and 97th percentile). Or "max" (default)
which will cut off the maximum value on either end.}

\item{print_v}{logical indicating whether or not to print a test heatmap}

\item{verbose_v}{logical indicating whether or not to print out messages}
}
\description{
Determine custom breaks for heatmap by excluding high values.
}
