% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtListBind.R
\name{dtListBind}
\alias{dtListBind}
\title{Data table list bind}
\usage{
dtListBind(list_lsdt, newName_v = "Name")
}
\arguments{
\item{list_lsdt}{named list of data.tables}

\item{newName_v}{character vector that will be the new column name}
}
\value{
single data.table with one more column that input data.tables and as 
many rows as all the input tables combined.
}
\description{
rbind a list of data.tables and also add the
names of the list elements as a new column in each dt.
}
