% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupCols.R
\name{dupCols}
\alias{dupCols}
\title{Duplicated Columns}
\usage{
dupCols(data_dt, rev_v = F, remove_v = T, verbose_v = F)
}
\arguments{
\item{data_dt}{input data.table with duplicated columns}

\item{rev_v}{logical. TRUE - keep last occurrence of duplicated name; FALSE (default) keep 1st occurrence}

\item{remove_v}{logical, TRUE - remove columns; FALSE - do not remove, just print column names to console (if verbose)}

\item{verbose_v}{logical. TRUE - print out duplicated columns. FALSE - do not print}
}
\value{
data.table same as data_dt, optionally with extra columns removed.
}
\description{
Identify duplicated column names in a data.table. Print them to console and optionally remove them from data.table.
Will keep the first (or last) occurrence of a column name and remove all subsequent occurrences.
}
\examples{
dat <- data.table("A" = 1:5, "B" = 6:10, "A" = 11:15)
dupCols(dat)
dupCols(dat, rev_v = T)
}
