% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBar.R
\name{facetStackedBar}
\alias{facetStackedBar}
\title{Stacked Bar}
\usage{
facetStackedBar(
  data_dt,
  xVar_v,
  facetVar_v,
  yVar_v = "value",
  fillVar_v = "variable",
  title_v = NULL,
  colors_v = NULL
)
}
\arguments{
\item{data_dt}{melted data.table}

\item{xVar_v}{variable to put on x-axis (should be 'Patients' or 'Time point')}

\item{facetVar_v}{variable to facet by (should be opposite of xVar_v, i.e. 'Time point' if xVar_v == 'Patients')}

\item{yVar_v}{variable to put on y-axis (should always be 'value')}

\item{fillVar_v}{variable to fill stacked bars b y (should always be 'variable')}

\item{title_v}{optional plot title}

\item{colors_v}{named color vector that matches values in fillVar_v (optional, but should be provided)}
}
\value{
ggplot
}
\description{
Stacked bar of different populations
}
\details{
This is made to make two different versions:
1. Facet by time, patients on the x-axis
2. Facet by patients, time on the x-axis
}
