% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdSubchunkify.R
\name{fdSubchunkify}
\alias{fdSubchunkify}
\title{FD Subchunkify}
\usage{
fdSubchunkify(g, fig_height = 7, fig_width = 5, id = NULL, title_v = "Title")
}
\arguments{
\item{g}{chunk}

\item{fig_height}{height of chunk}

\item{fig_width}{width of chunk}

\item{optional}{chunk ID}
}
\description{
Output a flexdashboard plotting chunk
Code taken from stack overflow and adapted.
https://stackoverflow.com/questions/15365829/dynamic-height-and-width-for-knitr-plots/47339394#47339394
References a blog post by Michael J Williams, but link is broken
Addition of ID from this post: https://stackoverflow.com/questions/61620768/rmarkdown-explicitly-specify-the-figure-size-of-a-plot-within-a-chunk
Had to modify the paste0() call for sub_chunk because r markdown was giving warnings about mismatched ticks for some reason.
Function to output variable figure sizes
}
