% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCommonCorrGenes.R
\name{findCommonCorrGenes}
\alias{findCommonCorrGenes}
\title{Find Common Corr Genes}
\usage{
findCommonCorrGenes(
  corrData_lsdt,
  corrCol_v = "Corr",
  geneCol_v = "Gene",
  thresh_v = NULL,
  option_v = "count"
)
}
\arguments{
\item{corrData_lsdt}{list of correlation results tables from tcgaCorrelation()}

\item{corrCol_v}{colname of tables that holds correlation values. Should always be 'Corr'}

\item{geneCol_v}{colnames of tables that holds the gene names. Should always be 'Gene'}

\item{thresh_v}{correlation threshold to use before finding intersection (abs value)}

\item{option_v}{need a better name. Either 'count' or 'exact'}
}
\value{
vector of common genes
}
\description{
Find genes common between corr tables
}
