% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMatch.R
\name{findMatch}
\alias{findMatch}
\title{Find match}
\usage{
findMatch(counts1_v, counts2_v, checks_v)
}
\arguments{
\item{counts1_v}{same scale as the values in checks_v}

\item{counts2_v}{scale that will find equivalent}

\item{checks_v}{values to test}
}
\description{
Given two matching vectors of different scales (e.g. raw and cpm),
find out which values map with each other, e.g. which cpm corresponds to a raw count of 10?
}
