% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixVSegments.R
\name{fixVSegments}
\alias{fixVSegments}
\title{Fix V Segments}
\usage{
fixVSegments(v_v)
}
\arguments{
\item{v_v}{character vector of segments to fix}
}
\description{
V segments are simplified during pipeline. Some tools need them converted back to
'standard' format. Add 'TRB' to beginning and add '-' when needed. (E.g. V132 -> TRBV13-2)
}
