% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGGLegend.R
\name{g_legend}
\alias{g_legend}
\title{Extract ggplot legend}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{a ggplot object with a legend}
}
\value{
a gtable object of the legend
}
\description{
Extract legend as separate gtable from ggplot object
}
