% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomMean.R
\name{geomMean}
\alias{geomMean}
\title{Geometric Mean}
\usage{
geomMean(counts_v)
}
\arguments{
\item{counts_v}{numeric vector of values}
}
\value{
geometric mean of vector
}
\description{
Find geometric mean of numeric vector
}
\examples{
set.seed(25)
foo <- sample(1:100, 5)
geomMean(foo)
}
