% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllListNames.R
\name{getAllListNames}
\alias{getAllListNames}
\title{Get All List Names}
\usage{
getAllListNames(list_ls, names_lsv = NULL, counter_v = 1)
}
\arguments{
\item{list_ls}{named list}

\item{names_lsv}{holds output. Recursively updated.}

\item{counter_v}{counter for recursion tracking}
}
\description{
Given a named list, compile names from each level until no more lists.
}
