% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColors.R
\name{getColors}
\alias{getColors}
\title{Get Colors}
\usage{
getColors(n_v)
}
\arguments{
\item{n_v}{number of colors to get}
}
\value{
vector of color hex codes
}
\description{
Get vector of length n_v from RColorBrewer
}
\details{
Uses Dark2 (8) first, then Set3 (12), and finally Pastel1 (9).
If n_v is greater than the total colors in those 3 sets, warning will be thrown.
}
