% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDims.R
\name{getDims}
\alias{getDims}
\title{Get Dimensions}
\usage{
getDims(n_v)
}
\arguments{
\item{n_v}{number of plots to get dimensions from}
}
\value{
vector of length two - c(row, col)
}
\description{
Get row/col dimensions for ggarrange
}
