% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreatmentComparisons.R
\name{getPairwiseComparisons}
\alias{getPairwiseComparisons}
\title{Get Pairwise Comparisons}
\usage{
getPairwiseComparisons(toCompare_v)
}
\arguments{
\item{toCompare_v}{vector to run comparisons for}
}
\value{
list.
}
\description{
Create a list of pairwise comparisons from an input vector

Returns a non-repeating list of pairs, so order of toCompare_v matters.
}
