% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrime.R
\name{isPrime}
\alias{isPrime}
\title{Is Prime?}
\usage{
isPrime(x, div_v = T, print_v = T)
}
\arguments{
\item{x}{number to test}

\item{div_v}{boolean determining if divisors should be output or not.}

\item{print_v}{boolean determining if answer should be printed to console}
}
\value{
list. 1st element - boolean of if x is prime or not. 2nd element - vector of divisors, if specified
}
\description{
Determine if a number is a prime number
}
\examples{
test_v <- c(2, 10, 17, 567, 5678)
sapply(test_v, isPrime)
sapply(test_v, function(x) isPrime(x, print_v = F, div_v = F))
}
