% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmPlot.R
\name{kmPlot}
\alias{kmPlot}
\title{Custom Kaplan Meier plot with median lines and legends}
\usage{
kmPlot(
  fit,
  model,
  colors_v = c("blue", "red"),
  labels_v = c(Low = "low", High = "up"),
  main_v = "KM Survival",
  xlab_v = "Overall Survival",
  ylab_v = "Prop. Survival",
  max_x,
  leg_y = c(med = 1, rec = 0.83, p = 0.7),
  fileName_v = NA
)
}
\arguments{
\item{fit}{"survfit" object created by survfit() function}

\item{model}{"survdiff" object created by survdiff() function}

\item{colors_v}{vector of colors to plot. Must be same length as number of divisions in fit}

\item{labels_v}{named vector of divisions in survival model. Names will be displayed on legend, values must match those of fit/model.}

\item{main_v}{title for plot}

\item{xlab_v}{label for X-axis. Default is "Overall Survival"}

\item{ylab_v}{label for y-axis. Default is "Prop. Survival"}

\item{max_x}{Max x-value from original data that created fit. Used for placing legends}

\item{leg_y}{named vector of legend y-axis placements. Names are 'med', 'p', and 'rec'. Legends will only be printed if they have a leg_y value.}

\item{fileName_v}{path to output file where plot will be printed. If NA, will print to stdout.}
}
\value{
plot to console of Kaplan Meier survival estimate
}
\description{
Specific Kaplan Meier plot
}
