% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelLog2.R
\name{label_log2}
\alias{label_log2}
\title{Label Log2 Ggplot}
\usage{
label_log2(x)
}
\arguments{
\item{x}{}
}
\description{
label log2 axes as 2^x.
Found on SO:
https://stackoverflow.com/questions/47879117/how-to-format-x-axis-tick-label-in-2x-format
}
