% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllListNames.R
\name{listNames}
\alias{listNames}
\title{Get List Names}
\usage{
listNames(list_ls, prefix_v = "")
}
\arguments{
\item{list_ls}{named list}

\item{prefix_v}{optional prefix to past in front of each element. Default is "" (blank)}
}
\value{
either a character vector or a matrix, depending on evenness of list depth
}
\description{
recursively get all the names of a nested list
}
