% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLib.R
\name{loadLib}
\alias{loadLib}
\title{Load Libraries}
\usage{
loadLib(lib_v, silent_v = T)
}
\arguments{
\item{lib_v}{character vector containing names of libraries to load}

\item{silent_v}{boolean determining if library load messages should be printed or not.}
}
\value{
load libraries and (optionally) print any messages to console
}
\description{
Load vector of libraries
}
\examples{
loadLib(c("data.table", "ggplot2", "dplyr"))
}
