% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggThemes.R
\name{massive_label}
\alias{massive_label}
\title{Massive Label Theme}
\usage{
massive_label(multiplier_v = 1)
}
\arguments{
\item{multiplier_v}{numeric. multiple text sizes by this factor.}
}
\description{
Same as big_label() theme, but even LARGER text and also y-axis labels are angled 45
}
