% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcFindAllMarkers.R
\name{mcFindAllMarkers}
\alias{mcFindAllMarkers}
\title{Run FindAllMarkers() on multiple cores}
\usage{
mcFindAllMarkers(obj, nCore_v, onlyPos_v = F, test_v = "MAST", latent_v)
}
\arguments{
\item{obj}{seurat object. Make sure clusters of interest are set as Idents()}

\item{nCore_v}{number of cores to split on.}

\item{onlyPos_v}{logical, passed to only.pos argument in FindMarkers.
default F. TRUE - only output positive log2FC markers
           FALSE - output positive and negative markers}

\item{test_v}{which test to use. Passed to test.use in FindMarkers}

\item{latent_v}{vector of latent variables to account for in DE testing.}
}
\value{
data.table of markers
}
\description{
Adapted from https://gist.github.com/diazdc/1735102c243cd16acb1b1f3fd09a26e1
https://github.com/diazdc/seurat-extensions/blob/master/R/cluster_markers.R
Instead of running FindAllMarkers, run FindMarkers() in parallel for each cluster.
}
