% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAndFillCols.R
\name{mergeAndFillCols}
\alias{mergeAndFillCols}
\title{Merge and Fill Columns}
\usage{
mergeAndFillCols(data_dt, toMerge_v, outName_v = "merged",
  search_lsv = NULL, replace_lsv = NULL)
}
\arguments{
\item{data_dt}{data.table with values}

\item{toMerge_v}{vector of column names that should be merged}

\item{outName_v}{name of the resulting merged column. Default is 'merged'.}

\item{search_lsv}{list of values to search for within the columns}

\item{replace_lsv}{list of values to replace. Each element must match with search_lsv}
}
\description{
Takes >= 2 columns with overlapping, incomplete data and merges them into 1 column.
}
