% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDTs.R
\name{mergeDTs}
\alias{mergeDTs}
\title{Merge many data.tables together}
\usage{
mergeDTs(data_lsdt, mergeCol_v, keepCol_v = NULL, ...)
}
\arguments{
\item{data_lsdt}{list of data.tables to merge}

\item{mergeCol_v}{which column from all of the data.tables to use to merge}

\item{keepCol_v}{which column from all of the data.tables to use as the column of interest. If NULL, use all columns}

\item{...}{extra parameters passed to merge. So far all = T and sort = F unless changed here.}
}
\value{
data.table with ncol == length(data_lsdt) + 1. Column names are names of list, or defaults to V1, V2,...
}
\description{
Take many data.tables and merge on and ID column, extracting a single column from each data.table as the column of interest
}
