% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhead.R
\name{mhead}
\alias{mhead}
\title{Special head for table-like objects}
\usage{
mhead(data_df, n = 5)
}
\arguments{
\item{data_df}{any data.table, data.frame, or matrix}

\item{n}{number of rows and columns to subset by. Default is 5}
}
\value{
prints to console data_df[1:n,1:n]
}
\description{
Show first n rows and columns of data.table/data.frame/matrix. Good for tables with many columns
}
\examples{
data_mat <- matrix(1:10000, nrow = 100)
mhead(data_mat)
mhead(data_mat, 10)
}
