% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdir.R
\name{mkdir}
\alias{mkdir}
\title{Creates new directory in which to write files}
\usage{
mkdir(baseDir_v = NULL, newDir_v, wd_v = F)
}
\arguments{
\item{baseDir_v}{Character string. Relative or absolute path to directory that will be parent of newly created directory. 
If set to NULL (default), then will use working directory.}

\item{newDir_v}{Character string. Name of new directory.}

\item{wd_v}{boolean values determining if new directory should be set as working directory (T) or not (F).}
}
\value{
Character string of path to new directory. 
Makes directory in file system. 
If wd_v == TRUE, also sets as working directory.
}
\description{
Given a base directory and string, will check if specified directory exits, and make it if not. 
Can also choose to set as working directory.
}
\examples{
mkdir("~/", "mkdir_test")
setwd("~/test/")
mkdir(newDir_v = "anotherTest")
mkdir("~/", "wdTest", wd_v = T)
myDir_v <- mkdir("~/", "finalTest")
}
