% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelP.R
\name{modelP}
\alias{modelP}
\title{Extract p-value from model}
\usage{
modelP(model, round_v = T, verbose_v = F)
}
\arguments{
\item{model}{model object of class 'lm' for linear model, classes 'lm' and 'aov' for ANOVA.}

\item{round_v}{logical. TRUE - round p-value to 3 decimals; FALSE - do not round.}

\item{verbose_v}{logical. TRUE - print model type; FALSE - don't print}
}
\value{
numeric vector of p-value
}
\description{
Extract the p-value estimate from either a linear model (lm) or ANOVA (aov) object.
}
\examples{
set.seed(1)
x <- sample(1:100, size = 10)
y <- sample(1:100, size = 10)
my_lm <- lm(y ~ x)
modelP(my_lm)
}
