% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSave.R
\name{multiSave}
\alias{multiSave}
\title{Multi Save}
\usage{
multiSave(
  obj_ls,
  file_v,
  height_v = 7,
  width_v = 7,
  res_v = 600,
  which_v = c("pdf", "png", "svg")
)
}
\arguments{
\item{obj_ls}{list of objects to print}

\item{file_v}{file path to output. Can have any file suffix - suffix is controlled by which_v}

\item{height_v}{height of file}

\item{width_v}{width of file}

\item{res_v}{resolution (for png)}

\item{which_v}{vector containing any combo of 'pdf', 'png', and 'svg', indicating what types to save.}
}
\description{
Save image(s) in multiple file types
}
