% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvFiles.R
\name{mvFiles}
\alias{mvFiles}
\title{Move a batch of files to a new directory}
\usage{
mvFiles(origDir_v, newDir_v, pattern_v = "*\\\\.log")
}
\arguments{
\item{origDir_v}{path to original directory containing the files}

\item{newDir_v}{one of two. (1) path to new directory where files should be moved to. (2) character string of new sub-directory of origDir_v to put files into.}

\item{pattern_v}{some sort of regex to uniquely identify files to move.}
}
\value{
moves files at system level
}
\description{
Originally developed to move venn diagram log output. Can be used to move any unique set of files to new directory.
}
