% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCorChart.R
\name{myCorChart}
\alias{myCorChart}
\title{My Correlation Chart}
\usage{
myCorChart(
  R,
  histogram = TRUE,
  method = c("pearson", "kendall", "spearman"),
  pch = 20,
  main = "Correlation"
)
}
\arguments{
\item{R}{data for the x axis, can take matrix, vector, or timeseries}

\item{histogram}{TRUE/FALSE whether or not to display a histogram}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman", can be abbreviated.}

\item{pch}{plotting symbol to use. see 'pch values' under ?pch}

\item{main}{optional title for plot}
}
\description{
Copy of PerformanceAnalytics::chart.Correlation (v2.0.4) with a few adjustments
}
\details{
PerformanceAnalytics::chart.Correlation() has a '...' argument to pass arguments to pairs(), but it doesn't work.
Needed to change the point type and add a title so copied the function and added that. Everything else is the same.
}
