% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKable.R
\name{myKable}
\alias{myKable}
\title{My Kable}
\usage{
myKable(data, format_v = "html", width_v = 100, position_v = "center", ...)
}
\arguments{
\item{data}{data.table/frame/matrix for plotting}

\item{format_v}{argument to kable's `format`. Usually 'html' or 'latex', see ?knitr for others.}

\item{position_v}{given to kable_styling position argument. Center is default. float_left, right, etc.}

\item{...}{extra arguments passed to knitr::kable}

\item{width}{number used as percentage of document width}
}
\value{
kable object
}
\description{
Function with a few common parameters I use already coded in
}
\details{
Use "float_left" and "right" in conjunction to put multiple tables on the same line
}
