% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myT.R
\name{myT}
\alias{myT}
\title{My Transpose
Transpose data. Make the first column of original data be the colnames
of the new data and the colnames of original data be either the 
rownames or the 1st column of new data (depends on output type)}
\usage{
myT(data_dt, newName_v, toClass_v = "numeric", outType_v = "dt")
}
\arguments{
\item{data_dt}{data to transpose}

\item{newName_v}{name for new 1st column of transposed data (only used if dt is output type)}

\item{toClass_v}{which class to convert output to (only 'numeric' and 'character' work right now, NULL to avoid.)}

\item{outType_v}{which type of object to return (dt for data.table, df for data.frame, mat for matrix)}
}
\description{
My Transpose
Transpose data. Make the first column of original data be the colnames
of the new data and the colnames of original data be either the 
rownames or the 1st column of new data (depends on output type)
}
