% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTableGrob.R
\name{myTableGrob}
\alias{myTableGrob}
\title{Create custom table grob with title}
\usage{
myTableGrob(data_dt, title_v, fontsize_v = 14)
}
\arguments{
\item{data_dt}{Data.table that the grob will be made out of}

\item{title_v}{Title for display}

\item{fontsize_v}{Fontsize for title. Default is 14 (goes well with my_theme)}
}
\value{
gtable object
}
\description{
Creates table grob in format that is most common for my usage.
}
\examples{
data_dt <- data.table("A" = LETTERS[1:5], "B" = 1:5)
myGrob <- myTableGrob(data_dt, title_v = "Test")
}
