% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTableSummary.R
\name{myTableSummary}
\alias{myTableSummary}
\title{My Table Summary}
\usage{
myTableSummary(data_dt, cols_v, addSum_v = T, verbose_v = F)
}
\arguments{
\item{data_dt}{data to summarize}

\item{cols_v}{column name(s) from data_dt that need to be summarized. Can only be 1, 2, or 3 columns (3 not recommended)}

\item{addSum_v}{logical if totals should be added at the bottom.}
}
\value{
data.table
}
\description{
Wrapper for table() that makes outputs nicer and the way I like them
}
\details{
Works for 1 to 3 columns, although 3 not recommended.
}
