% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myWriteLines.R
\name{myWriteLines}
\alias{myWriteLines}
\title{My Write Lines}
\usage{
myWriteLines(file_v, data_lsv, sep_v = " ")
}
\arguments{
\item{file_v}{path to file}

\item{data_lsv}{list of vectors}

\item{sep_v}{Character vector used to separate elements of a vector. If NULL, each element of each vector will have its own line}
}
\value{
no output to stdout, just writes to file.
}
\description{
Write a list of vectors to an output file. Each line is one element of list.
}
\examples{
testOut_lsv <- list(rep(0, 5), LETTERS[1:10], "This is a test")
myWriteLines("~/Desktop/multiLineTest.txt", testOut_lsv, sep = NULL)
myWriteLines("~/Desktop/singleLineTest.txt", testOut_lsv)
}
