% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nChooseK.R
\name{nChooseK}
\alias{nChooseK}
\title{n Choose k}
\usage{
nChooseK(n_v, k_v)
}
\arguments{
\item{n_v}{number of possible elements to create combinations}

\item{k_v}{size of combinations (e.g. 3 means all unique combinations of 3 elements from n_v)}
}
\value{
numeric vector listing the total number of combinations
}
\description{
Given 'n' elements and combinations of size 'k', find number of unique combinations.
}
\examples{
set.seed(1)
n_v <- sample(1:50, size = 5)
k1_v <- 3; k2_v <- 5
nChooseK(n_v[1], k1_v)
nChooseK(n_v[1], k2_v)
nChooseK(n_v, k1_v)
nChooseK(n_v, k2_v)
}
