% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notice.R
\name{notice}
\alias{notice}
\title{Print easy-to-find statement}
\usage{
notice(statement_v)
}
\arguments{
\item{statement_v}{Some sort of character vector. Can be a paste() call, an sprintf message, etc.}
}
\value{
print to console
}
\description{
Print a specified statement after a large ASCII "LOOK" statement.
}
\examples{
notice("This is a test")
notice(paste("Another", "test", sep = "_"))
}
