% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot Densities}
\usage{
plotDensity(
  density_ls,
  colors_ls = brewer.pal(8, "Dark2"),
  main_v = "",
  x_v = "",
  names_v = NULL,
  lwd_v = 2
)
}
\arguments{
\item{density_ls}{list of either densities or vectors}

\item{colors_ls}{list of colors one for each element of density_ls}

\item{main_v}{character vector for title}

\item{x_v}{character vector for x-axis}

\item{names_v}{names to plot to legend. If NULL, will use names of density_ls.}

\item{lwd_v}{line width}
}
\value{
print density plot to console
}
\description{
Plot multiple density functions on same graph
}
\examples{
set.seed(1)
d_ls <- list("A" = rnorm(1000), "B" = rnorm(1000, mean = 2, sd = 1), "C" = rnorm(1000, mean = -1, sd = 0.5))
plotDensity(d_ls)
}
