% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profitMarginCalculator.R
\name{profitMarginCalculator}
\alias{profitMarginCalculator}
\title{Profit Margin Calculator}
\usage{
profitMarginCalculator(price_v, margin_v = NULL, shipping_v = NULL)
}
\arguments{
\item{price_v}{numeric vector of the price of the product in dollars}

\item{margin_v}{numeric value from 0.0 to 1.0 that is the profit margin in a decimal percentage (e.g. 15% == 0.15)}

\item{shipping_v}{numeric value of shipping cost in dollars.}
}
\value{
numeric vector of suggested price
}
\description{
Calculate appropriate price depending on desired margin and potential shipping costs.
}
\details{
Use supplied info to calculate the appropriate price for a product.
}
