% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAllExcel.R
\name{readAllExcel}
\alias{readAllExcel}
\title{Read All Excel}
\usage{
readAllExcel(file_v, ...)
}
\arguments{
\item{file_v}{path to excel file. Can be '.xlsx' or '.xls'}

\item{...}{extra options to pass to readxl::read_excel}
}
\value{
list of data.tables
}
\description{
Read in all of the sheets in an excel workbook as a list of data.tables
}
