% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnSessionInfo.R
\name{returnSessionInfo}
\alias{returnSessionInfo}
\title{Output session information}
\usage{
returnSessionInfo(outDir_v = NULL, loadDirs_v = NULL, args_lsv = NULL)
}
\arguments{
\item{outDir_v}{optional directory to write information to. Default is to print to stdout}

\item{loadDirs_v}{vector of directory paths that have been sourced in script. Note, if using the utility function
sourceDir, these will be the same paths.}

\item{args_lsv}{optional list of arguments used as input. Could be from optparse (easiest), or can hand-make a list of arguments.}
}
\value{
multiple lines of text printed to stdout or written to file
}
\description{
Print general R session info, current date/time, and loaded git repo versions to stdout or file. Note
this uses sessionInfo() from base R, but an argument could be made to use session_info() from devtools.
}
\examples{
returnSessionInfo()
returnSessionInfo("~/Desktop/")
}
