% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revComp.R
\name{revComp}
\alias{revComp}
\title{Reverse Complement
Given vector of strings of ATCG, provide reverse complement of each string}
\usage{
revComp(string_v)
}
\arguments{
\item{string_v}{vector of strings to be reverse complemented}
}
\value{
Vector of same length, reverse complemented
}
\description{
Reverse Complement
Given vector of strings of ATCG, provide reverse complement of each string
}
