% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDupGenes.R
\name{rmDupGenes}
\alias{rmDupGenes}
\title{Remove Duplicate Genes}
\usage{
rmDupGenes(data_dt, idCol_v = "ID", symbolCol_v = "Symbol",
  method_v = "max")
}
\arguments{
\item{data_dt}{data.table of gene expression.}

\item{idCol_v}{don't think this is needed actually}

\item{symbolCol_v}{Name of column that contains gene identifiers}

\item{method_v}{One of "max" (default) or "mean". Max means that the gene with max count will be taken, "mean" means will average them all.}
}
\description{
Can't make data.frames with duplicate rownames. Have to remove duplicates.
}
