% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNARow.R
\name{rmNARow}
\alias{rmNARow}
\title{Remove rows that have NA in certain columns}
\usage{
rmNARow(data_dt, cols_v = NULL, extract_v = F)
}
\arguments{
\item{data_dt}{data.table or data.frame to have rows removed}

\item{cols_v}{columns to check for NA in. If NULL (default) will check all columns}

\item{extract_v}{logical, TRUE - output the NA rows; FALSE (default) - output the input table with NA rows removed}
}
\value{
data.table either containing the offending rows (extract_v == T), or the input table with those rows removed
}
\description{
Remove all rows that have an NA in at least one of the specified columns.
}
\examples{
data_df <- data.frame("A" = 1:5, "B" = c(1,2,NA,4,5), "C" = c(NA,2:5), "D" = c(1:4,NA))
rmNARow(data_dt = data_df, cols_v = "A")
rmNARow(data_dt = data_df, cols_v = c("A", "B"))
rmNARow(data_dt = data_df)
rmNARow(data_dt = data_df, extract_v = T)
}
