% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFGSEA.R
\name{runFGSEA}
\alias{runFGSEA}
\title{FGSEA Wrapper}
\usage{
runFGSEA(
  data_df,
  rankCol_v = "avg_log2FC",
  pathways_v,
  minSize_v = 1,
  maxSize_v = nrow(data_df),
  seed_v = 42
)
}
\arguments{
\item{data_df}{table of data to run. must have log2FC column and column named 'genes'}

\item{rankCol_v}{column name of data_df to use to rank. Default is avg_log2FC}

\item{pathways_v}{list of gene pathways to test against}

\item{minSize_v}{passed to fgsea. minimum size of gene set to test}

\item{maxSize_v}{passed to fgsea. Maximum size of gene set to test}

\item{seed_v}{set for reproducibility}
}
\description{
Run fgsea and wrangle output
}
