% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shippingCalculator.R
\name{shippingCalculator}
\alias{shippingCalculator}
\title{Calculate Shipping}
\usage{
shippingCalculator(weight_v = 0, shop_v)
}
\arguments{
\item{weight_v}{numeric value of product weight in pounds.}

\item{shop_v}{character vector. either "PlantBoy" or "PlantShop"}
}
\value{
vector of price
}
\description{
Calculate shipping cost based on weight for plantboy
}
