% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Allows users to source all .R files in a given directory}
\usage{
sourceDir(dirPath_v, recurse_v = TRUE, verbose_v = F)
}
\arguments{
\item{dirPath_v}{a vector of strings, each corresponding to a directory containing .R files you would like sourced}

\item{recurse_v}{a boolean allowing the user to specify whether the sourcing should dive into subdirectories. 
Unlike list.files(), the default is true}

\item{verbose_v}{a boolean determining if sourced files should be printed.}
}
\value{
The list of full paths to .R scripts being sourced
}
\description{
Enables convenient sourcing of multiple .R files simultaneously. 
Optionally, sourcing can occur recursively in subdirectories.
}
\examples{
sourceDir(dirPath_v=c("/path/to/UtilityFunctions", "/path/to/AbundanceFunctions"))
}
