% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitChar.R
\name{splitChar}
\alias{splitChar}
\title{Split character string}
\usage{
splitChar(vector_v)
}
\arguments{
\item{vector_v}{any character vector that needs to be separated into its component parts}
}
\value{
another vector of length == nchar(input)
}
\description{
Split character string into vector where length(output) == nchar(input)
}
\examples{
foo <- "ABCDE"
splitChar(foo)
}
