% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitComma.R
\name{splitComma}
\alias{splitComma}
\title{Split a string along the commas in the string}
\usage{
splitComma(string_v)
}
\arguments{
\item{String}{that will be split}
}
\value{
Vector of strings. One element for each section of string that was previously divided by a comma
}
\description{
Just a concise way to split a string on the commas. Returning a vector of strings instead.
}
\examples{
myString_v <- c("This,is,a,string")
splitComma(myString_v)
}
