% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitInputNumbers.R
\name{splitInputNumbers}
\alias{splitInputNumbers}
\title{Split Input Numbers}
\usage{
splitInputNumbers(char_v, split_v = ",")
}
\arguments{
\item{char_v}{Character vector that is input}

\item{split_v}{Delimiter used to separate numbers.}
}
\description{
When list of numbers are provided as a character (usually as command-line argument, or user text input),
split them into a numeric vector. Also supports splitting "4:8" into 4, 5, 6, 7, 8.
}
\examples{
char1_v <- "1,2,3,4,5"
char2_v <- "1:5"
char3_v <- "1:5,9,10,20:25"
splitInputNumbers(char1_v)
splitInputNumbers(char2_v)
splitInputNumbers(char3_v)
}
