% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBar.R
\name{stackedBar}
\alias{stackedBar}
\title{Stacked Bar}
\usage{
stackedBar(
  data_dt,
  xVar_v = "Time point",
  patientCol_v = "Patients",
  popCol_v = "variable",
  patient_v,
  pops_v = NULL,
  yVar_v = "value",
  fillVar_v = "variable",
  title_v = NULL,
  colors_v = NULL
)
}
\arguments{
\item{data_dt}{melted data.table}

\item{xVar_v}{variable to put on x-axis (should be 'Patients' or 'Time point')}

\item{patientCol_v}{column name specifying different patients. Should be 'Patients' always.}

\item{popCol_v}{column name specifying different populations. Should always be 'variable'}

\item{patient_v}{specific value in patientCol_v to subset for}

\item{pops_v}{specific populations to display, if NULL (default), will use all available}

\item{yVar_v}{variable to put on y-axis (should always be 'value')}

\item{fillVar_v}{variable to fill stacked bars b y (should always be 'variable')}

\item{title_v}{optional plot title}

\item{colors_v}{named color vector that matches values in fillVar_v (optional, but should be provided)}
}
\value{
ggplot
}
\description{
Stacked bar of different populations
}
