% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackName.R
\name{stackName}
\alias{stackName}
\title{Stack Name for Plot}
\usage{
stackName(x_v, delim_v = "_")
}
\arguments{
\item{x_v}{long string with delimiter}

\item{delim_v}{delimiter, usually _ for msigdb names}
}
\value{
same string, but a few of delim_v replaced with newline
}
\description{
Given a string with a lot of a certain delimiter, replace that delimiter with \n 
in certain spots. Made for MsigDB GO pathway names, but use where applicable.
}
