% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{standardPlots}
\alias{standardPlots}
\title{Standard Seurat Plots}
\usage{
standardPlots(seurat_obj, reduction_v, clustCol_v, res_v, name_v)
}
\arguments{
\item{seurat_obj}{seurat object to use for plotting}

\item{reduction_v}{name of reduction to use for UMAP - usually "umap" or "harmony"}

\item{clustCol_v}{column name of seurat clusters to plot}

\item{res_v}{resolution used for clusters}

\item{name_v}{name to add to plot titles}
}
\value{
list of ggplot objects
}
\description{
Make some quick plots for unintegrated, harmony, or other objects
}
