% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaCorrelation.R
\name{tcgaCorrelation}
\alias{tcgaCorrelation}
\title{TCGA Correlation}
\usage{
tcgaCorrelation(
  data_dt,
  name_v = NULL,
  refGene_v = "IL27",
  corType_v = "pearson",
  sampleCol_v = "sample",
  sigThresh_v = NULL,
  corrThresh_v = NULL
)
}
\arguments{
\item{data_dt}{TCGA cancer data (see details)}

\item{name_v}{optional name for cancer type}

\item{refGene_v}{which gene to correlate others to}

\item{corType_v}{type of correlation to run}

\item{sampleCol_v}{column name that refers to unique rows. Should be sample}

\item{sigThresh_v}{optional value to subset for significant results}

\item{corrThresh_v}{optional value to subset for high correlations (abs value)}
}
\value{
tbd
}
\description{
Get correlation between ref gene and all other genes
}
\details{
data_dt has rows as samples and columns as gene expression and metadata
meta columns are theoretically variable, but the LAST column should end in "time"
and the sample column should be 'sample'
}
