% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thirds.R
\name{thirds}
\alias{thirds}
\title{Thirds-based vector division}
\usage{
thirds(x_v, na.rm = T)
}
\arguments{
\item{x_v}{numeric vector}

\item{na.rm}{TRUE (default) remove NAs. FALSE - don't remove. returns all NA}
}
\value{
vector containing value of min, bottom 3rd, median, top 3rd, and max.
}
\description{
Instead of dividing a vector into quartiles, divide into thirds and return values at those points.
}
\examples{
set.seed(1)
a <- sample(1:1000, size = 100)
b <- sample(1:100000, size = 1000)
thirds(a)
thirds(b)
}
