% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrh.rUtils-package.R
\docType{package}
\name{wrh.rUtils-package}
\alias{wrh.rUtils}
\alias{wrh.rUtils-package}
\title{wrh.rUtils: Miscellaneous R Utility Functions}
\description{
Just a collection of different functions I've written to make my life easier.
}
\author{
\strong{Maintainer}: Wes Horton \email{wrhorton592@gmail.com}

}
\keyword{internal}
