% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSVorExcel.R
\name{writeCSVorExcel}
\alias{writeCSVorExcel}
\title{Write CSV or Excel}
\usage{
writeCSVorExcel(obj, file_v, quote_v = F, rowNames_v = F)
}
\arguments{
\item{obj}{either a data.table, data.frame, or matrix, or a list of one of those}

\item{file_v}{path to file to write out. Can be either .csv or .xlsx}

\item{quote_v}{logical. passed to quote argument of write.csv}

\item{rowNames_v}{logical. passed to row.names argument of write.csv}
}
\value{
write out file.
}
\description{
Will write out object according to file_v's extension
}
\details{
Will write .csv or .xlsx depending on the extension found in file_v
Currently can only write a single table to csv. Additional functionality
Would be to write a list of tables to a directory of csv's instead of an excel file.
}
