% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatedBar.R
\name{annotatedBar}
\alias{annotatedBar}
\title{Annotated Bar Plot}
\usage{
annotatedBar(
  data_dt = NULL,
  plot_gg = NULL,
  x_v,
  y_v = NULL,
  stat_v = "count",
  position_v = "stack",
  yLab_v = "count",
  fill_v = NULL,
  fillColors_v = NULL,
  keepXAxis_v = F,
  title_v = NULL,
  annot_lsv = NULL,
  annotColors_lsv,
  testTime_v = F,
  backgroundCol_v = "white"
)
}
\arguments{
\item{data_dt}{melted data table for plotting (see details. Must supply this or plot_gg)}

\item{plot_gg}{ggplot to annotate (see details. Must supply this or data_dt)}

\item{x_v}{name of x-axis variable (usually Sample_ID, Treatment, etc.). Must be column in data_dt}

\item{y_v}{name of y-axis variable (onnly used if stat_v = "identity")}

\item{stat_v}{argument to 'stat' parameter of geom_bar. 'count' is default and does not use y_v. "identity" requires y_v to be set}

\item{position_v}{argument to 'position' parameter of geom_bar. 'stack' (default) is counts; 'fill' turns to percentage by filling out of 1.}

\item{yLab_v}{name of y-axis label.}

\item{fill_v}{name of fill variable. Must be column in data_dt}

\item{fillColors_v}{optional named color vector. Values are colors, names are values of data_dt[[fill_v]]}

\item{keepXAxis_v}{logical indicating if x-axis labels should be included. Default is false. Should only be used if there aren't too many x-axis values}

\item{title_v}{optional plot title}

\item{annot_lsv}{list of annotations to add below the plot. e.g. annot_lsv = list("outName1" = "colName1", "outName2" = "colName2")}

\item{annotColors_lsv}{list of colors for annotations. list element names are same as annot_lsv names; list elements are named color vectors, names are values of annot_lsv colNames in data_dt}

\item{testTime_v}{logical indicating whether to see how long it takes to plot}

\item{backgroundCol_v}{sent to panel.background element of ggplot theme. Needs to be white for a specific use case where I want to plot a percentage using fill_v = "fill", but I only want a subset of the identities to be shown.}
}
\value{
list of gg objects output by ggarrange. Combo plot, data only, annotations only.
}
\description{
Standard bar-plot with heatmap-like x-axis annotations
}
\details{
make a standard ggplot bar plot with extra annotations. Can provide a pre-existing plot instead of data.
If a plot is provided, the x_v variable is still required, along with annot_lsv and annotColors_lsv. 
Not required if plot is given: stat_v, position_v, fill_v, fillColors_v
Note that data_dt overrules plot_gg. If both are provided, plot_gg is ignored
}
