% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDFT.R
\name{convertDFT}
\alias{convertDFT}
\title{Convert between data.table and data.frame}
\usage{
convertDFT(data_dft, col_v = NA, newName_v = "V1", rmCol_v = T)
}
\arguments{
\item{data_dft}{data in either data.table or data.frame format (can also be a matrix)}

\item{col_v}{character or numeric vector. if converting from dt to df, column name or index of which column to use as row.names.
NA (default) will use 1st column; NULL will not add rownames}

\item{newName_v}{character vector. if converting from df/mat to dt, what to name new column. (default is "V1")}

\item{rmCol_v}{boolean value indicating whether to remove the column used to make the rownames from the output table (T) or to leave it (F)}
}
\value{
either a data.table or data.frame (opposite class of input)
}
\description{
Change data.tables into data.frames with specified row.names or
data.frames/matrices into data.tables, copying over row.names as the first column.
If data.frame/matrix doesn't have row.names, then no columns will be added to data.table
}
\examples{
# Data
my_df <- data.frame("A" = 1:10, "B" = LETTERS[1:10], "C" = letters[11:20])
my_df2 <- my_df; rownames(my_df2) <- paste0("Row", 1:10)
my_mat <- as.matrix(my_df); my_mat2 <- as.matrix(my_df2)
my_dt <- data.table("AA" = 10:1, "BB" = LETTERS[5:14], "CC" = letters[20:11])
convertDFT(data_dft = my_df)
convertDFT(data_dft = my_df2)
convertDFT(data_dft = my_df2, newName_v = "Test")
convertDFT(data_dft = my_mat2, newName_v = "MatTest")
convertDFT(data_dft = my_dt)
convertDFT(data_dft = my_dt, col_v = NULL)
convertDFT(data_dft = my_dt, col_v = "BB")
convertDFT(data_dft = my_dt, col_v = 3, rmCol_v = F)
}
