% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPCs.R
\name{findPCs}
\alias{findPCs}
\title{Find maximum recommended PC to use.}
\usage{
findPCs(seurat_obj, drName_v = "pca")
}
\arguments{
\item{seurat_obj}{A standard seurat object that has a "pca" dimensional reduction}

\item{drName_v}{Name of pca dimensional reduction. Default is 'pca', which is Seurat default. 
If you create a PCA on specific genes and label it something else, provide that label name 
here (e.g. ccs_pca) for a PCA on cell cycle genes.}
}
\value{
A single numeric value with the recommended PC cut-off
}
\description{
Using a seurat object with PCA dimensional reduction run on it, determine the highest PC that 
still provides information about variance between cells.
}
