% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutliers.R
\name{getOutliers}
\alias{getOutliers}
\title{Find Outliers}
\usage{
getOutliers(data_dt, col_v, type_v = c("mean", "mad"), by_v = NULL)
}
\arguments{
\item{data_dt}{data.table containing distribution to search for}

\item{col_v}{column name that contains data}

\item{type_v}{vector indicating which outlier method(s) to use. Can be either or both of "mean" and "mad"}

\item{by_v}{column name indicating grouping variable. Default is no grouping.}
}
\description{
find outliers in distributions using MAD or mean +- 2 SD
}
