% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPairComparisons.R
\name{getPairwiseComparisons}
\alias{getPairwiseComparisons}
\title{Get Pairwise Comparisons}
\usage{
getPairwiseComparisons(toCompare_v, unique_v = T)
}
\arguments{
\item{toCompare_v}{vector to run comparisons for}

\item{unique_v}{logical indicating if only unique comparisons should be run}
}
\value{
list.
}
\description{
Create a list of pairwise comparisons from an input vector

unique_v = T will return a non-repeating list of pairs, so order of toCompare_v matters. unique_v = F will return all pairs
}
