% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDT.R
\name{listToDT}
\alias{listToDT}
\title{Convert List to data.table}
\usage{
listToDT(input_lsv)
}
\arguments{
\item{input_lsv}{list of vectors. Can be of varying lengths and classes}
}
\value{
data.table object where each column corresponds to an element of the input list.
Number of rows in output data.table is equal to the length of the longest vector in input list.
}
\description{
Transform a list of vectors into a data.table. 
Vectors can be of differing lengths (shorter vectors padded with NA).
Each list element becomes a column in the new table.
}
\examples{
list1_ls <- list("A" = 1:5, "B" = LETTERS[1:10], "C" = letters[11:16])
list2_ls <- list("2" = 10:1, "5" = LETTERS[1:10])
list1_dt <- listToDT(list1_ls)
list2_dt <- listToDT(list2_ls)
}
