% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKable.R
\name{myKable}
\alias{myKable}
\title{My Kable}
\usage{
myKable(data, width_v)
}
\arguments{
\item{data}{data.table/frame/matrix for plotting}

\item{width}{number used as percentage of document width}
}
\value{
kable object
}
\description{
Function with a few common parameters I use already coded in
}
