% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profitMarginCalculator.R
\name{profitMarginCalculator}
\alias{profitMarginCalculator}
\title{Profit Margin Calculator}
\usage{
profitMarginCalculator(
  price_v,
  margin_v = NULL,
  shippingWeight_v = 0,
  shop_v,
  includeShippingInMargin_v = F
)
}
\arguments{
\item{price_v}{numeric vector of the price of the product in dollars}

\item{margin_v}{numeric value from 0.0 to 1.0 that is the profit margin in a decimal percentage (e.g. 15% == 0.15)}

\item{shippingWeight_v}{numeric value of shipping weight in pounds.}

\item{shop_v}{either "PlantBoy" or "PlantShop" indicating which supplier}

\item{includeShippingInMargin_v}{logical to add shipping cost before calculating margin}
}
\value{
numeric vector of suggested price
}
\description{
Calculate appropriate price depending on desired margin and potential shipping costs.
}
\details{
Use supplied info to calculate the appropriate price for a product.
}
