% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSurvival.R
\name{simpleSurvival}
\alias{simpleSurvival}
\title{Simple Survival}
\usage{
simpleSurvival(
  data_dt,
  exprCol_v,
  timeCol_v = "OS.time",
  deathCol_v = "OS",
  method_v = "quartile",
  outputFile_v = NULL,
  name_v = NULL
)
}
\arguments{
\item{data_dt}{data.table with combined clinical and expression data}

\item{exprCol_v}{column that contains expression information to use as a group divider.}

\item{timeCol_v}{column that indicates survival time. Originally was days_to_death, may now be OS.time or DSS.time}

\item{deathCol_v}{column that indicates death (vital) status. "DECEASED" == 1; "LIVING" == 0}

\item{method_v}{which method to use to divide expression groups. quartile (default) will take the upper and lower quartiles. Median will split in half on the median.}

\item{outputFile_v}{optional path to an output location if saving file is desired}

\item{name_v}{optional name to add to the title. Usually the title of the dataset used as input.}
}
\value{
plot
}
\description{
Calculate Kaplan Meier survival estimate and plot. Split groups based on expression.
}
\details{
This function is specifically made for data that has been downloaded from the TCGA hub by the UCSCXenaTools package.
[here](https://xenabrowser.net/datapages/?dataset=survival%2FBRCA_survival.txt&host=https%3A%2F%2Ftcga.xenahubs.net&removeHub=https%3A%2F%2Fxena.treehouse.gi.ucsc.edu%3A443) is an example of the BRCA
dataset that is accessed by this tool. There are multiple survival classifications in these datasets:
  1. OS: overall survival
  2. DSS: disease-specific survival
  3. DFI: disease-free interval
  4. PFI: progression-free interval
}
